Use dfa to speed up translating UTF-8 into code point
This dfa, available from the internet, has the reputation of being the
fastest general translator.  It does not allow Perl extended UTF-8.
This commit changes to use it at the beginning of our translator,
modifying it to accept surrogates, and all 4-byte UTF-8.  If a probem
occurs, it drops down into our regular translator to handle
warnings and errors and 5-and-up-byte Perl extended.