hints/hpux.sh: HP-UX mbrlen() and mbrtowc() don't work
In spite of there being man pages for these, the #include file doesn't
define the mbstate_t type which is required for a parameter to these
functions.

Perhaps the Configure probe could be enhanced so it doesn't return
defined unless these can be successfully compiled, but for now use the
hints file.